Fix travis build
(cherry picked from commit b6e46325d3d7ce92885c7658881480057b6eb420)

+++ tests/src/core/testqgscoordinatereferencesystem.cpp
@@ -369,7 +369,7 @@ void TestQgsCoordinateReferenceSystem::createFromWktUnknown()
   // When used with proj < 6, a lossy conversion to proj string is used
   QCOMPARE( crs.toWkt(), wkt );
 #endif
-  QCOMPARE( crs.srsid(), USER_CRS_START_ID + 1 );
+  QCOMPARE( crs.srsid(), static_cast< long >( USER_CRS_START_ID + 1 ) );
   QCOMPARE( crs.authid(), QStringLiteral( "USER:100001" ) );
   QCOMPARE( crs.mapUnits(), QgsUnitTypes::DistanceMeters );
